requirements.txt
discord_py_interactions>=5.13.2
python-dotenv>=1.1.1
python_dateutil>=2.9.0.post0
Requests>=2.32.5
